게임 접속까지 가능하도록 변경.

+++ Thieves/Server/Refactored_Server/Thieves_Server_Refactored/Thieves_Server_Refactored/Session.h
@@ -29,7 +29,7 @@ public:
 
 		this->SendRaw(&packetData, sizeof(T));
 
-		DEBUG_LOG("[" << packetId << " Packet ] Send \n");
+		DEBUG_LOG("[" << (int)packetId << " Packet ] Send \n");
 	}
 
 	template <typename T>
